requirements.txt
absl-py==0.10.0
antlr4-python3-runtime==4.8
anykeystore==0.2
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.2.0
audeer==1.16.0
audiofile==1.0.0
audioread==2.1.9
audtorch==0.6.4
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.6
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cxxfilt==0.3.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
easydict==1.9
et-xmlfile==1.0.1
Flask==1.1.2
future==0.18.2
gast==0.2.2
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==1.1.1
grpcio==1.33.1
h5py==2.10.0
hupper==1.10.3
hydra-core==1.1.1
idna==2.10
importlib-metadata==2.0.0
importlib-resources==5.2.2
iniconfig==1.1.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
joblib==0.17.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa==0.8.1
llvmlite==0.34.0
Markdown==3.3.2
MarkupSafe==1.1.1
matplotlib==3.3.2
networkx==2.5
numba==0.51.2
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
omegaconf==2.1.1
opencv-python==4.5.4.60
openpyxl==3.0.6
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.5
PasteDeploy==2.1.1
pbkdf2==1.3
Pillow==8.0.1
pipreqs==0.4.11
plaster==1.0
plaster-pastedeploy==0.7
plotly==5.5.0
pluggy==1.0.0
pooch==1.5.1
protobuf==3.13.0
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pynndescent==0.5.1
pyparsing==2.4.7
pyramid==2.0
pyramid-mailer==0.15.1
pytest==6.2.5
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
repoze.sendmail==4.4.1
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.11.2
six==1.15.0
SoundFile==0.10.3.post1
sox==1.4.1
SQLAlchemy==1.4.25
tabulate==0.8.9
tb-nightly==1.14.0a20190301
tenacity==8.0.1
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
threadpoolctl==2.1.0
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
tqdm==4.50.2
transaction==3.0.1
translationstring==1.4
typing-extensions==3.7.4.3
umap-learn==0.4.6
urllib3==1.25.11
velruse==1.1.1
venusian==3.0.0
WebOb==1.8.7
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
wtforms-recaptcha==0.3.2
xlrd==2.0.1
yarg==0.1.9
zipp==3.3.1
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
